intelligent-qa-service/requirements.txt
# intelligent-qa-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
mysql-connector-python==8.2.0
numpy==1.24.3
scikit-learn==1.3.0
asyncio-mqtt==0.16.1
python-multipart==0.0.6
